pyproject.toml
[project]
name = "agent-cell-poc"
version = "0.1.0"
description = "Agent Cell PoC: Network Security & Observability"
requires-python = ">=3.12"
dependencies = [
    "confluent-kafka>=2.6.0",
    "psycopg[binary]>=3.2.0",
    "pgvector>=0.3.0",
    "anthropic>=0.42.0",
    "typer[all]>=0.15.0",
    "rich>=13.9.0",
    "orjson>=3.10.0",
    "numpy>=2.0.0",
    "sentence-transformers>=3.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
agentcell = "src.cli.main:app"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
